pyproject.toml
[tool.poetry]
name = "client"
version = "0.1.0"
description = ""
authors = ["Noam Dolovich <noamdolovich@microsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
urllib3 = "^1.26.13"
opentelemetry-api = "^1.15.0"
opentelemetry-sdk = "^1.15.0"
opentelemetry-exporter-jaeger = "^1.15.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.8.0"
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
